feat: rewrite in python
This adds a small python module for working with the input data and the
ontologies. The data can be converted from CI artifacts such as JUnit
XML and requirements tracing files from cargo-requirements. The
resulting files are then checked by a dedicated command to provide
a Boolean certification status for the CI. The detailed certification
report can be generated using the provided Sphinx extension.

An example is provided in .gitlab-ci.yml and the `doc` sub-directory.